Create the UI for the opening screen
Create the user interface for the `MessagesViewController`, which is the
first view controller shown when the user opens the app. Inside of it,
add a woodgrain background, two `UIButton`s that say "Play" and "Learn",
and a `UILabel` that says "speedPoker". Although no items should be
interactive, they will be in the future.